+++ CHANGES.md
@@ -2,6 +2,10 @@
 
 ### ? - ?
 
+##### Breaking Changes :mega:
+
+- Removed `tilesLoadingLowPriority`, `tilesLoadingMediumPriority`, and `tilesLoadingHighPriority` from `ViewUpdateResult`. Use `workerThreadTileLoadQueueLength` and `mainThreadTileLoadQueueLength` instead.
+
 ##### Additions :tada:
 
 - Added `getOrientedBoundingBoxFromBoundingVolume` to the `Cesium3DTilesSelection` namespace.
@@ -15,6 +19,10 @@
 ##### Fixes :wrench:
 
 - Fixed a bug that caused the `center` field of `AxisAlignedBox` to be incorrect.
+
+##### Fixes :wrench:
+
+- Fixed a bug that caused the main thread to sometimes load low-priority tiles before high-priority ones. This could result in much longer waits than necessary for a tileset's appropriate level-of-detail to be shown.
 
 ### v0.22.1 - 2023-03-06